#31: Add new value converters: InverseBoolConverter, NotNullConverter, and NullConverter
- Introduced InverseBoolConverter to invert boolean values.
- Added NotNullConverter to return true for non-null values and false otherwise.
- Implemented NullConverter to return true for null values and false otherwise.
- Removed the previous ValueConverters.cs file to streamline the converter structure.
