--- OtChaim.Presentation.MAUI/Converters/ValueConverters.cs
@@ -1,59 +1,0 @@
-using System.Globalization;
-
-namespace OtChaim.Presentation.MAUI.Converters;
-
-/// <summary>
-/// Converter that returns true if the value is not null, false otherwise.
-/// </summary>
-public class NotNullConverter : IValueConverter
-{
-    public object? Convert(object? value, Type targetType, object? parameter, CultureInfo culture)
-    {
-        return value != null;
-    }
-
-    public object? ConvertBack(object? value, Type targetType, object? parameter, CultureInfo culture)
-    {
-        throw new NotImplementedException();
-    }
-}
-
-/// <summary>
-/// Converter that returns true if the value is null, false otherwise.
-/// </summary>
-public class NullConverter : IValueConverter
-{
-    public object? Convert(object? value, Type targetType, object? parameter, CultureInfo culture)
-    {
-        return value == null;
-    }
-
-    public object? ConvertBack(object? value, Type targetType, object? parameter, CultureInfo culture)
-    {
-        throw new NotImplementedException();
-    }
-}
-
-/// <summary>
-/// Converter that inverts a boolean value.
-/// </summary>
-public class InverseBoolConverter : IValueConverter
-{
-    public object? Convert(object? value, Type targetType, object? parameter, CultureInfo culture)
-    {
-        if (value is bool boolValue)
-        {
-            return !boolValue;
-        }
-        return value;
-    }
-
-    public object? ConvertBack(object? value, Type targetType, object? parameter, CultureInfo culture)
-    {
-        if (value is bool boolValue)
-        {
-            return !boolValue;
-        }
-        return value;
-    }
-}
